+++ test/formatters/installers/taplo.bash
@@ -1,0 +1,1 @@
+npm install -g @taplo/cli